requirements.txt
aiohttp
imgbb
Pillow
requests
python-dotenv
pyperclip